requirements.txt
python-telegram-bot==20.7
ccxt==4.1.92
aiohttp==3.9.1
aiosqlite==0.19.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0

